+++ R/get_emtrends.R
@@ -16,6 +16,7 @@
 get_emtrends <- function(model,
                          trend = NULL,
                          by = NULL,
+                         keep_iterations = FALSE,
                          verbose = TRUE,
                          ...) {
   # check if available
@@ -33,10 +34,19 @@ get_emtrends <- function(model,
     ...
   )
 
+  # for Bayesian model, keep iterations
+  if (insight::model_info(model)$is_bayesian) {
+    attr(estimated, "posterior_draws") <- insight::get_parameters(estimated)
+  } else {
+    keep_iterations <- FALSE
+  }
+
   attr(estimated, "trend") <- my_args$trend
   attr(estimated, "at") <- my_args$by
   attr(estimated, "by") <- my_args$by
   attr(estimated, "coef_name") <- "Slope"
+  attr(estimated, "transform") <- TRUE
+  attr(estimated, "keep_iterations") <- keep_iterations
 
   estimated
 }
@@ -83,7 +93,6 @@ get_emtrends <- function(model,
 
 # Formatting ===============================================================
 
-
 .format_emmeans_slopes <- function(model, estimated, ci, ...) {
   # Summarize and clean
   if (insight::model_info(model)$is_bayesian) {
@@ -105,5 +114,8 @@ get_emtrends <- function(model,
   trends <- datawizard::data_rename(trends, select = c(Slope = "Coefficient"))
 
   # Restore factor levels
-  datawizard::data_restoretype(trends, insight::get_data(model, verbose = FALSE))
+  out <- datawizard::data_restoretype(trends, insight::get_data(model, verbose = FALSE))
+
+  # add posterior draws?
+  .add_posterior_draws_emmeans(attributes(estimated), out)
 }