+++ R/get_marginaltrends.R
@@ -12,6 +12,8 @@ get_marginaltrends <- function(model,
                                by = NULL,
                                ci = 0.95,
                                p_adjust = "none",
+                               transform = NULL,
+                               keep_iterations = FALSE,
                                verbose = TRUE,
                                ...) {
   # check if available
@@ -47,7 +49,6 @@ get_marginaltrends <- function(model,
     datagrid_info <- attributes(datagrid)
   }
 
-
   # Second step: prepare arguments for marginaleffects ------------------------
   # ---------------------------------------------------------------------------
 
@@ -77,12 +78,29 @@ get_marginaltrends <- function(model,
     dots
   ))
 
-
   # Third step: compute marginal slopes ---------------------------------------
   # ---------------------------------------------------------------------------
 
   # Compute stuff
   estimated <- suppressWarnings(do.call(marginaleffects::avg_slopes, fun_args))
+
+  # Fourth step: back-transform response --------------------------------------
+  # ---------------------------------------------------------------------------
+
+  # transform reponse?
+  if (isTRUE(transform)) {
+    trans_fun <- insight::get_transformation(model, verbose = FALSE)$inverse
+  } else {
+    trans_fun <- transform
+  }
+  # if we have back-transformation, do that, but remove standard errors
+  # these are no longer correct
+  if (!is.null(trans_fun)) {
+    estimated$estimate <- trans_fun(estimated$estimate)
+    estimated$conf.low <- trans_fun(estimated$conf.low)
+    estimated$conf.high <- trans_fun(estimated$conf.high)
+    estimated$std.error <- NULL
+  }
 
 
   # Last step: Save information in attributes  --------------------------------
@@ -98,7 +116,9 @@ get_marginaltrends <- function(model,
         datagrid = datagrid,
         coef_name = "Slope",
         p_adjust = p_adjust,
-        ci = ci
+        ci = ci,
+        transform = !is.null(transform),
+        keep_iterations = keep_iterations
       )
     )
   )